ENH: Allow limiting of log to a passed revision range
So that the "rabbitvcs log" command can be invoked after e.g. a "git pull" and shows exactly those commits that have been pulled.

Note: I did not want to define and pass through a new command-line parsing, so this is passed via a RABBITVCS_REVISION_RANGE environment variable.
Limited to Git and Subversion so far.